featrue: 总model及相关api设计
